fix: reload confirmation alert and stay focused after saving
